==> src/Notebooks/4_Modeling/3.2_XGBoost_Model_Functions.ipynb <==
import pandas as pd
import numpy as np
import xgboost

def identification_0H(reading):
    threshold = 0.5
    
    model = xgboost.Booster()
    model.load_model('../../data/model/Identificacao_0H.json')
    
    df = pd.DataFrame(
        data=[reading],
        columns=['Mes', 'Dia', 'Local', 'Precipitacao', 'PrecSum']
    )
    
    data = xgboost.DMatrix(data=df)
    result = bst.predict(data)
    
    result = True if result > threshold else False
    
    return result

bst = xgboost.Booster()
bst.load_model('../../data/model/Identificacao_0H.json')

df = pd.DataFrame(data=[[3, 1, 2, 2.2, 64]], columns=['Mes', 'Dia', 'Local', 'Precipitacao', 'PrecSum'])
data = xgboost.DMatrix(data=df)
print(bst.predict(data))
